+++ app/mainmenu.h
@@ -22,6 +22,7 @@ signals:
   void exit();
   void toggleTrayIcon();
   void waveformToggled(bool enabled);
+  void playlistHeaderToggled(bool enabled);
   void openPlaybackLog();
   void openShortcuts();
   void openEqualizer();