+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Dynamic range analyzer (playlist context menu or track info dialog), in Foobar's text format
 - Editable keyboard shortcuts
+- Floating playlist columns header (default off, enabled in settings)
 
 ### Fixed